Added scoin support to discord bot and website (#52)
* Added support for setting coin values on discord bot.

* Working web leaderboard for scoin.

* Proper scoreboard of scoin.

* Added scoin permission and checks for single permission.

* Got permissions working with discord for allowing setting coin amounts.

* Added permissions to web standings.

* Regenerated and removed credential.

* Added credentials file for oauth to prevent leaking credentials.

* Added warning about not using postgres to bot config.

* Fixed formatting.

* Cleaned up files.

+++ engfrosh_site/discord_bot_manager/models.py
@@ -40,5 +40,5 @@ class ScavChannel(models.Model):
 
 
 class Role(models.Model):
-    role_id = models.IntegerField("Discord Role ID", primary_key=True)
+    role_id = models.PositiveBigIntegerField("Discord Role ID", primary_key=True)
     group_id = models.OneToOneField(Group, CASCADE)
